minor refactor
rename package to move withdrawn appeal classes into it
Rename handlers based on the event name
refactor test method to reuse across both test classes

Add unit test class for new handler
